\docType{package}
\name{MedTst}
\alias{MedTst}
\title{Creamos distintas medidas de testeo específicas para distintos percentiles y métodos alternativos en la Venta Directa}
\usage{
DUPlot()
}
\description{
Con este paquete podemos categorizar por niveles de asertividad, generar la mejor elección posible y brindar inputs para generar niveles de servicio e inventario.
}
\author{
\itemize{
  \item Jenny Patiño \email{jmpatinoa@unal.edu.com}
  \item Joselyn Perez \email{joselyn.c.perez.aguirre@gmail.com}
  \item Anthony Castro \email{acastroa@unp.edu.pe}
}
}
